refactor: split gnrsqldata.py into thematic sub-modules
Split the monolithic gnrsqldata.py (~2646 lines) into 4 focused modules:
- gnrsqldata_compiler.py: SqlCompiledQuery, SqlQueryCompiler
- gnrsqldata_query.py: SqlQuery, SqlDataResolver
- gnrsqldata_selection.py: SqlSelection
- gnrsqldata_record.py: SqlRecord, SqlRecordBag, SqlRelatedRecordResolver, SqlRelatedSelectionResolver

gnrsqldata.py is now a facade that re-exports all classes for backward compatibility.
All imports are top-level, no circular dependencies.

All 329 existing tests pass unchanged.
